optimizations

+++ lib/hittable/hittable.h
@@ -10,7 +10,7 @@ class hit_record {
  public:
     point3 p;
     vec3 normal;
-    double t;
+    float t;
     bool front_face;
     shared_ptr<material> mat;
