+++ lib/objects/sphere.h
@@ -6,7 +6,7 @@
 
 class sphere : public hittable {
  public:
-    sphere(point3 _center, double _radius, shared_ptr<material> _material)
+    sphere(point3 _center, float _radius, shared_ptr<material> _material)
         : center(_center), radius(_radius), mat(_material) {
     }
 
@@ -43,7 +43,7 @@ class sphere : public hittable {
 
  private:
     point3 center;
-    double radius;
+    float radius;
     shared_ptr<material> mat;
 };
